split index tests
